+++ src/BladeRouteGenerator.php
@@ -1,31 +1,29 @@
 <?php
 
-namespace Tightenco\Ziggy;
+namespace Tighten\Ziggy;
 
-use Tightenco\Ziggy\Output\MergeScript;
-use Tightenco\Ziggy\Output\Script;
+use Tighten\Ziggy\Output\MergeScript;
+use Tighten\Ziggy\Output\Script;
 
 class BladeRouteGenerator
 {
     public static $generated;
 
-    public function generate($group = null, $nonce = null)
+    public function generate($group = null, string $nonce = null): string
     {
         $ziggy = new Ziggy($group);
 
-        $nonce = $nonce ? ' nonce="' . $nonce . '"' : '';
+        $nonce = $nonce ? " nonce=\"{$nonce}\"" : '';
 
         if (static::$generated) {
             return (string) $this->generateMergeJavascript($ziggy, $nonce);
         }
 
-        $function = $this->getRouteFunction();
-
         static::$generated = true;
 
         $output = config('ziggy.output.script', Script::class);
 
-        return (string) new $output($ziggy, $function, $nonce);
+        return (string) new $output($ziggy, $this->getRouteFunction(), $nonce);
     }
 
     private function generateMergeJavascript(Ziggy $ziggy, $nonce)
@@ -37,6 +35,6 @@ class BladeRouteGenerator
 
     private function getRouteFunction()
     {
-        return config('ziggy.skip-route-function') ? '' : file_get_contents(__DIR__ . '/../dist/index.js');
+        return config('ziggy.skip-route-function') ? '' : file_get_contents(__DIR__ . '/../dist/route.umd.js');
     }
 }