+++ src/ZiggyServiceProvider.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tightenco\Ziggy;
+namespace Tighten\Ziggy;
 
 use Illuminate\Support\Facades\Event;
 use Illuminate\Support\ServiceProvider;
@@ -14,9 +14,7 @@ class ZiggyServiceProvider extends ServiceProvider
         if ($this->app->resolved('blade.compiler')) {
             $this->registerDirective($this->app['blade.compiler']);
         } else {
-            $this->app->afterResolving('blade.compiler', function (BladeCompiler $bladeCompiler) {
-                $this->registerDirective($bladeCompiler);
-            });
+            $this->app->afterResolving('blade.compiler', fn (BladeCompiler $blade) => $this->registerDirective($blade));
         }
 
         Event::listen(RequestReceived::class, function () {
@@ -28,10 +26,8 @@ class ZiggyServiceProvider extends ServiceProvider
         }
     }
 
-    protected function registerDirective(BladeCompiler $bladeCompiler)
+    protected function registerDirective(BladeCompiler $blade)
     {
-        $bladeCompiler->directive('routes', function ($group) {
-            return "<?php echo app('" . BladeRouteGenerator::class . "')->generate({$group}); ?>";
-        });
+        $blade->directive('routes', fn ($group) => "<?php echo app('" . BladeRouteGenerator::class . "')->generate({$group}); ?>");
     }
 }