+++ src/js/browser.js
@@ -1,0 +1,1 @@
+export { route as default } from './index.js';