+++ src/js/index.d.ts
@@ -173,9 +173,9 @@ interface Router {
  * Ziggy's route helper.
  */
 // Called with no arguments - returns a Router instance
-export default function route(): Router;
+export function route(): Router;
 // Called with a route name and optional additional arguments - returns a URL string
-export default function route<T extends RouteName>(
+export function route<T extends RouteName>(
     name: T,
     params?: RouteParams<T> | undefined,
     absolute?: boolean,
@@ -188,9 +188,21 @@ export default function route<T extends RouteName>(
     config?: Config,
 ): string;
 // Called with configuration arguments only - returns a configured Router instance
-export default function route(
+export function route(
     name: undefined,
     params: undefined,
     absolute?: boolean,
     config?: Config,
 ): Router;
+
+/**
+ * Ziggy's Vue plugin.
+ */
+export const ZiggyVue: {
+    install(app: any, options?: Config): void;
+};
+
+/**
+ * Ziggy's React hook.
+ */
+export function useRoute(defaultConfig?: Config): typeof route;