+++ tests/Unit/CommandRouteGeneratorTest.php
@@ -5,7 +5,7 @@ namespace Tests\Unit;
 use Illuminate\Support\Facades\Artisan;
 use Illuminate\Support\Facades\URL;
 use Tests\TestCase;
-use Tightenco\Ziggy\Output\File;
+use Tighten\Ziggy\Output\File;
 
 class CommandRouteGeneratorTest extends TestCase
 {
@@ -166,10 +166,6 @@ class CommandRouteGeneratorTest extends TestCase
     /** @test */
     public function can_generate_dts_file_with_scoped_bindings()
     {
-        if (! $this->laravelVersion(7)) {
-            $this->markTestSkipped('Requires Laravel >=7');
-        }
-
         app('router')->get('posts', $this->noop())->name('posts.index');
         app('router')->get('posts/{post}/comments/{comment:uuid}', PostCommentController::class)->name('postComments.show');
         app('router')->post('posts/{post}/comments', PostCommentController::class)->name('postComments.store');
@@ -198,7 +194,7 @@ class CommandRouteGeneratorTest extends TestCase
         Artisan::call('ziggy:generate', ['--types-only' => true]);
 
         $this->assertFileExists(base_path('resources/js/ziggy.d.ts'));
-        $this->assertFileNotExists(base_path('resources/js/ziggy.js'));
+        $this->assertFileDoesNotExist(base_path('resources/js/ziggy.js'));
     }
 
     /** @test */
@@ -212,7 +208,7 @@ class CommandRouteGeneratorTest extends TestCase
         Artisan::call('ziggy:generate', ['--types-only' => true]);
 
         $this->assertFileExists(base_path('resources/js/custom.d.ts'));
-        $this->assertFileNotExists(base_path('resources/js/ziggy.d.ts'));
+        $this->assertFileDoesNotExist(base_path('resources/js/ziggy.d.ts'));
     }
 
     /** @test */